fix(mongodb): break circular dependency and key dynamicOptions by model URI
- Create state.ts to hold dynamicOptions and mongoSampleQueries
- Key dynamicOptions by model URI to prevent multi-tab clobbering
- Import allCollectionMethods from keywords.ts instead of hard-coded list
- Remove flawed per-line quote validator (bracket validator handles strings)
- Remove duplicate MONGO_VALIDATION_OWNER constant
